AMLS-4079 Refactored DataCacheConnectorMigrator
